Tasks in Cards
Now displays an object of type Task in cards using Collection View, added icons for buttons, compose modal that uses regex to detect title, date, and time and add it to the tasks list to display in cards view. An object of type Task only has String variables for now, will change later.