Notifications + Extras
Swiping now also undoes cards, when they are done.
Cards are marked as done with a green checkmark.
Done button is now removed.
Tag matching is now case insensitive.
Notifications 30 minutes before the task are now scheduled, if possible and activated in settings.
Tapping outside the typing area now dismisses the keyboard in the New Tag and Compose Task views.